rename plots/core to plots/time_dependent
