+++ astroplan/plots/__init__.py
@@ -1,2 +1,2 @@
-from .core import *
+from .time_dependent import *
 from .sky import *
